requirements.txt
torch
monai
tqdm
pylidc
scikit-image
scipy
Pillow
einops
ml-collections
timm
nibabel
pyradiomics


